stt/requirementsLinux.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cuda-toolkit==13.0.2
evdev==1.9.3
filelock==3.24.3
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai-whisper==20250625
pillow==12.1.0
pycparser==3.0
pynput==1.8.1
python-xlib==0.33
regex==2026.2.19
requests==2.32.5
scipy==1.17.1
setuptools==78.1.0
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.12.0.dev20260303+cu130
torchvision==0.26.0.dev20260303+cu130
tqdm==4.67.3
triton==3.6.0+git9844da95
typing_extensions==4.15.0
urllib3==2.6.3
